modified Athour

+++ database/migrations/2023_11_07_124302_create_reviews_table.php
@@ -14,9 +14,9 @@ return new class extends Migration
         Schema::create('reviews', function (Blueprint $table) {
             $table->id();
             $table->text('text')->nullable();
-            $table->unsignedInteger('rate');
-            $table->unsignedInteger('book_id');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('rate');
+            $table->unsignedBigInteger('book_id');
+            $table->unsignedBigInteger('user_id');
             $table->timestamps();
 
             $table->foreign('book_id')->references('id')->on('books');
